enh: add vectors for boosting and return, refactor some internals
