+++ setup.cfg
@@ -50,8 +50,11 @@ fromdecay =
     zfit >=0.12.0
     zfit-physics >= 0.6
     decaylanguage >= 0.12.0     # not required but everyone using this feature will likely use DecayLanguage
+vector =
+    vector >= 1.0.0
 test =
     %(fromdecay)s
+    %(vector)s
     awkward>=1.0,<2.0
     coverage
     flaky
